Decreased window size and increased framerate
The samller windows size has a few effects:

1. the frequency calculation will be less accurate, because less
iterations of the binary search will be effective.
2. 'mixed signals', measures at the start or end of a note where half
the signal is one thing and the other half is something else will have
less impact, since they will be on the screen for less time.

The higher frame rate also looks really nice.